refactor(types): hocProps 타입 분리
