+++ src/types/components/hocProps.ts
@@ -1,0 +1,10 @@
+export interface WithAuthProps {
+    requireAuth?: boolean;
+    requireUnauth?: boolean;
+    requireRegistration?: boolean;
+}
+
+export interface WithLoadingProps {
+    isLoading?: boolean;
+    LoadingComponent?: React.ComponentType;
+}
